weather/pyproject.toml
[project]
name = "weather"
version = "0.1.0"
description = "A weather mcp server"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.5",
    "mcp[cli]>=1.27.0",
    "python-dotenv>=1.2.2",
]

[tool.pyright]
venvPath = "."
venv = ".venv"
